+++ app/src/main/java/com/gezimos/katapult/util/PrefsManager.kt
@@ -233,6 +233,14 @@ class PrefsManager(context: Context) {
         get() = boolPref(KEY_SHOW_BATTERY, false)
         set(value) = prefs.edit().putBoolean(KEY_SHOW_BATTERY, value).apply()
 
+    var showAlarm: Boolean
+        get() = boolPref(KEY_SHOW_ALARM, true)
+        set(value) = prefs.edit().putBoolean(KEY_SHOW_ALARM, value).apply()
+
+    var showWeather: Boolean
+        get() = boolPref(KEY_SHOW_WEATHER, false)
+        set(value) = prefs.edit().putBoolean(KEY_SHOW_WEATHER, value).apply()
+
     var roundedIcons: Boolean
         get() = boolPref(KEY_ROUNDED_ICONS, false)
         set(value) = prefs.edit().putBoolean(KEY_ROUNDED_ICONS, value).apply()
@@ -542,6 +550,8 @@ class PrefsManager(context: Context) {
         private const val KEY_CLOCK_FORMAT = "clock_format"
         private const val KEY_DATE_FORMAT = "date_format"
         private const val KEY_SHOW_BATTERY = "show_battery"
+        private const val KEY_SHOW_ALARM = "show_alarm"
+        private const val KEY_SHOW_WEATHER = "show_weather"
         private const val KEY_ROUNDED_ICONS = "rounded_icons"
         private const val KEY_DARK_MODE = "dark_mode"
         private const val KEY_HIDE_STATUS_BAR = "hide_status_bar"